requirements.txt
altgraph==0.15
asgiref==3.2.5
certifi==2018.4.16
click==6.7
Django==3.0.4
django-crispy-forms==1.9.0
django-admin-autocomplete-filter==0.5
django-filter==2.2.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
macholib==1.9
markdown2==2.3.8
MarkupSafe==1.1.1
modulegraph==0.16
Pillow==6.2.1
psycopg2-binary==2.8.5
py2app==0.14
python-decouple==3.3
python-http-client==3.2.6
pytz==2019.3
sendgrid==6.2.0
six==1.14.0
sqlparse==0.3.0
virtualenv==16.6.0
